Changed getDollarBalance to double return

+++ app/src/main/java/edu/uc/bitcoinparkpay/service/IAddressService.java
@@ -1,7 +1,5 @@
 package edu.uc.bitcoinparkpay.service;
 
-import java.math.BigDecimal;
-
 /**
  * Created by Greg on 6/18/2015.
  */
@@ -11,7 +9,7 @@ public interface IAddressService {
      *
      * @return the account balance in USD.
      */
-    public BigDecimal getDollarBalance( String label ) throws Exception;
+    public double getDollarBalance( String label ) throws Exception;
 
     /**
      Sends bitcoin to another address; checks for balance before calling AddressDAO.send
